Failed Factory Method Design Pattern: Flexibility and Extensibility
