+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/food/HolidayEnum.java
@@ -1,0 +1,8 @@
+package com.shatteredpixel.shatteredpixeldungeon.items.food;
+
+public enum HolidayEnum {
+    NONE,
+    EASTER, //TBD
+    HWEEN,//2nd week of october though first day of november
+    XMAS //3rd week of december through first week of january
+}
